pyproject.toml
[tool.ruff]
line-length = 88         # default characters
indent-width = 4
target-version = "py39"  # The minimum Python version to target
src = ["src", "scripts"]

[tool.ruff.lint]
select = ["E",   # pycodestyle errors
          "W",   # pycodestyle warnings
          "F",   # flake8
          "D",   # pydocstyle
          "I",   # isort
          # "RET", # flake8-return
          # "RUF", # ruff rules
        ]
ignore = ["E203", "E501", "D10", "D212", "D415"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-rA -x --doctest-modules --color=yes --cov=inspect_ai"
testpaths = ["tests"]
doctest_optionflags = ["NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

[tool.mypy]
warn_unused_ignores = true
no_implicit_reexport = true
strict_equality = true
warn_redundant_casts = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module="inspect_ai.*"
warn_return_any = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
extra_checks = true

[[tool.mypy.overrides]]
module = "pandas-stubs.*"
ignore_errors = true

[tool.poetry]
name = "mmrr"
version = "0.1.0"
description = "Multimodal Reference Relation Analysis"
authors = ["SInadumi <inazumi19@icloud.com>"]
readme = "README.md"
package-mode=true
packages = [
    { include = "mmrr", from = "src" },
    { include = "tools", from = "src/mmrr" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"

torch = [
    { platform = "darwin", version = "2.3.0", source = "PyPI" },
    { platform = "linux", url = "https://download.pytorch.org/whl/cu121/torch-2.3.0%2Bcu121-cp39-cp39-linux_x86_64.whl" },
]
torchvision =[
    { platform = "darwin", version = "0.18.0" },
    { platform = "linux", url = "https://download.pytorch.org/whl/cu121/torchvision-0.18.0%2Bcu121-cp39-cp39-linux_x86_64.whl"},
]
transformers = { version = "~4.42.3", extras = ["sentencepiece"] }
lightning = "~2.2.2"
tokenizers = "^0.19.1"
torchmetrics = "^1.1"
pandas = "^2.2.3"

rich = "^13.9.2"
wandb = "^0.16.0"

omegaconf = "^2.3"
hydra-core = "^1.3"
dataclasses-json = "^0.6.1"
typing-extensions = ">=4.4"
pydantic = "^2.9.2"

jaconv = ">0.3.4"
rhoknp = ">=1.5.1,<1.7.0"
kwja = "^2.4.0"
cohesion-tools = ">=0.5.7"

h5py = "^3.12.1"
[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
mypy = "^1.10.0"
pytest = "^8.2.0"
pytest-env = "^1.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
